requirements.txt
djangorestframework==3.13.1
drf-yasg ==1.21.3
django_rest_swagger ==2.2.0
djangorestframework-simplejwt==5.2.0